# Improve selection change (amongst others) performance.