--- a/editor/src/components/canvas/canvas-strategies/strategies/draw-to-insert-text-strategy.tsx
+++ b/editor/src/components/canvas/canvas-strategies/strategies/draw-to-insert-text-strategy.tsx
@@ -78,6 +78,7 @@ export const drawToInsertTextStrategy: MetaCanvasStrategy = (
 
         const textEditableAndHasText = MetadataUtils.targetTextEditableAndHasText(
           canvasState.startingMetadata,
+          canvasState.startingElementPathTree,
           targetParent.intendedParentPath,
         )