--- a/editor/src/components/canvas/canvas-strategies/strategies/reorder-slider-strategy.tsx
+++ b/editor/src/components/canvas/canvas-strategies/strategies/reorder-slider-strategy.tsx
@@ -33,7 +33,11 @@ export function reorderSliderStategy(
     canvasState.startingMetadata,
     target,
   )
-  const siblings = MetadataUtils.getSiblingsUnordered(canvasState.startingMetadata, target)
+  const siblings = MetadataUtils.getSiblingsUnordered(
+    canvasState.startingMetadata,
+    canvasState.startingElementPathTree,
+    target,
+  )
   const isAutoLayouted =
     MetadataUtils.isParentYogaLayoutedContainerAndElementParticipatesInLayout(
       target,