--- a/editor/src/components/canvas/canvas-strategies/strategies/reparent-as-static-strategy.tsx
+++ b/editor/src/components/canvas/canvas-strategies/strategies/reparent-as-static-strategy.tsx
@@ -150,8 +150,9 @@ function applyStaticReparent(
               canvasState.startingMetadata,
             ) ?? zeroCanvasRect
 
-          const siblingsOfTarget = MetadataUtils.getChildrenPathsUnordered(
+          const siblingsOfTarget = MetadataUtils.getChildrenPathsOrdered(
             canvasState.startingMetadata,
+            canvasState.startingElementPathTree,
             newParent.intendedParentPath,
           )