--- a/editor/src/components/canvas/canvas-strategies/strategies/reparent-helpers/reparent-property-strategies.ts
+++ b/editor/src/components/canvas/canvas-strategies/strategies/reparent-helpers/reparent-property-strategies.ts
@@ -32,6 +32,7 @@ import { cssNumber } from '../../../../inspector/common/css-utils'
 import { setProperty } from '../../../commands/set-property-command'
 import { mapDropNulls } from '../../../../../core/shared/array-utils'
 import * as EP from '../../../../../core/shared/element-path'
+import { ElementPathTrees } from '../../../../../core/shared/element-path-tree'
 
 type ReparentPropertyStrategyUnapplicableReason = string
 
@@ -52,7 +53,9 @@ interface ElementPathSnapshots {
 
 interface MetadataSnapshots {
   originalTargetMetadata: ElementInstanceMetadataMap
+  originalPathTrees: ElementPathTrees
   currentMetadata: ElementInstanceMetadataMap
+  currentPathTrees: ElementPathTrees
 }
 
 export const stripPinsConvertToVisualSize =
@@ -290,7 +293,11 @@ export const setZIndexOnPastedElement =
     metadata: MetadataSnapshots,
   ): ReparentPropertyStrategy =>
   () => {
-    const siblings = MetadataUtils.getChildrenUnordered(metadata.currentMetadata, targetParent)
+    const siblings = MetadataUtils.getChildrenOrdered(
+      metadata.currentMetadata,
+      metadata.currentPathTrees,
+      targetParent,
+    )
     const maximumZIndexOfOverlappingElements = mapDropNulls((sibling) => {
       return foldEither(
         () => null,