… editor/src/components/canvas/canvas-actions.spec.tsx
… editor/src/components/canvas/canvas-strategies/strategies/convert-to-absolute-and-move-strategy.spec.browser2.tsx
… editor/src/components/canvas/canvas-strategies/strategies/reparent-helpers/reparent-property-changes.ts
… editor/src/components/canvas/canvas-strategies/strategies/reparent-helpers/reparent-strategy-parent-lookup.ts
… editor/src/components/canvas/canvas-strategies/strategies/reparent-helpers/reparent-strategy-sibling-position-helpers.ts
… editor/src/components/canvas/canvas-strategies/strategies/set-flex-gap-strategy.tsx
… editor/src/components/canvas/canvas-strategies/strategies/set-padding-strategy.tsx
… editor/src/components/canvas/canvas-strategies/strategies/shared-absolute-resize-strategy-helpers.ts
… editor/src/components/canvas/canvas-strategies/strategies/shared-keyboard-strategy-helpers.ts
… editor/src/components/canvas/canvas-strategies/strategies/shared-move-strategies-helpers.ts
… editor/src/components/canvas/canvas-utils.ts
… editor/src/components/canvas/canvas.ts
… editor/src/components/canvas/commands/show-reorder-indicator-command.ts
… editor/src/components/canvas/commands/wrap-in-container-command.ts
… editor/src/components/canvas/controls/absolute-children-outline.tsx
… editor/src/components/canvas/controls/breadcrumb-trail.tsx
… editor/src/components/canvas/controls/guideline-helpers.ts
… editor/src/components/canvas/controls/insertion-plus-button.tsx
… editor/src/components/canvas/controls/new-canvas-controls.tsx
… editor/src/components/canvas/controls/reorder-slider-control.tsx
… editor/src/components/canvas/controls/select-mode/flex-gap-control.tsx
… editor/src/components/canvas/controls/select-mode/flex-resize-control.tsx
… editor/src/components/canvas/controls/select-mode/scene-label.tsx
… editor/src/components/canvas/controls/select-mode/select-mode-hooks.tsx
… editor/src/components/canvas/controls/select-mode/yoga-utils.ts
… editor/src/components/canvas/controls/text-edit-mode/text-edit-mode-hooks.tsx
… editor/src/components/canvas/controls/zero-sized-element-controls.tsx
… editor/src/components/canvas/gap-utils.ts
… editor/src/components/canvas/padding-utils.tsx
… editor/src/components/common/shared-strategies/convert-to-flex-strategy.ts