**Problem:**
Currently selection changes are still quite an expensive operation mostly because there is a change to the metadata and the inspector needs to be largely completely re-rendered.

As previously mentioned in https://github.com/concrete-utopia/utopia/pull/3755, it was observed that once `MetadataUtils.getChildrenOrdered` is made a lot faster, then it could replace `MetadataUtils.getChildrenUnordered` as that is currently performing a O(n) operation every time it is called.

**Fix:**
The primary part of this fix has just been to replace `getChildrenUnordered` with `getChildrenOrdered` and `getChildrenPathsUnordered` with `getChildrenPathsOrdered`, which necessitated a lot of threading instances of `ElementPathTrees` to the numerous locations where each of those were called.

A secondary change was made to remove the use of `ComplexMap` from `getElementWarningsInner` as that was causing a surprising 4-5ms of wastage during each call to the latter.

**Results:**
The top level result is that the mousedown handler goes from an average of 140ms to 120ms to process, which was measured by ignoring the first result and then taking the average of 3 subsequent measurements.

The cost change can be observed here comparing the before and then after profiles:
![before](https://github.com/concrete-utopia/utopia/assets/217400/ae2b066a-d764-4b7e-810c-d3cf9246031f)
![after](https://github.com/concrete-utopia/utopia/assets/217400/7148065e-8a71-401d-8f24-673f90c41fd3)

The entire `editorDispatch` call highlighted goes from around 19ms to around 10ms, because the `deriveState` call has massively shrunk as `getElementWarningsInner` almost disappears off the profile.

On the right hand side of the first profile the `getNameAndIconResult` can also be seen dominating that side, in the after case it has dropped to 1/10th of the profile time.

**Commit Details:**
- Removed `ComplexMap` from the return type of `getElementWarningsInner`.
- Replaced calls to `getChildrenUnordered` with `getChildrenOrdered`.
- Replaced calls to `getChildrenPathsUnordered` with `getChildrenPathsOrdered`.
- Added `reorderSubTreeChildren` which extracts out some special case logic so we can run it on a chunk of children, not necessarily on the entire listing of elements.
- Lots of functions now have 1 or more instances of `ElementPathTrees` added to their parameters, often to pass through into other functions where they've similarly had them added.